Issue #177 - Cast translation ids in list to Integer instead of Long

+++ GAE/src/org/waterforpeople/mapping/app/web/rest/TranslationRestService.java
@@ -348,7 +348,7 @@ public class TranslationRestService {
 	public Map<String, RestStatusDto> bulkDeleteTranslationById(
 			@RequestBody TranslationBulkDeletePayload payLoad) {
 		final Map<String, RestStatusDto> response = new HashMap<String, RestStatusDto>();
-		final List<Long> tIds = payLoad
+		final List<Integer> tIds = payLoad
 				.getTranslations();
 		RestStatusDto statusDto = null;
 		Translation t = null;
